Show stability score
