+++ lib/screens/health/health_card_view_model.dart
@@ -1,0 +1,9 @@
+
+class HealthCardViewModel {
+  HealthCardViewModel(this.value, this.change);
+  HealthCardViewModel.changeFromValueBefore(this.value, double valueBefore)
+    : change = value != null && valueBefore != null ? value - valueBefore : null;
+
+  final double value;
+  final double change;
+}
