+++ test/util/stability_score_tests.dart
@@ -15,17 +15,17 @@ void main() {
       final abnormal = _givenSessionGroup(SessionStatus.abnormal, 1);
       final crashed = _givenSessionGroup(SessionStatus.crashed, 1);
 
-      var session = Sessions([], [healthy]);
-      expect(session.stabilityScore(), isNull);
+      var sessions = Sessions([], [healthy]);
+      expect(sessions.stabilityScore(), isNull);
 
-      session = Sessions([], [healthy, errored]);
-      expect(session.stabilityScore(), isNull);
+      sessions = Sessions([], [healthy, errored]);
+      expect(sessions.stabilityScore(), isNull);
 
-      session = Sessions([], [healthy, errored, abnormal]);
-      expect(session.stabilityScore(), isNull);
+      sessions = Sessions([], [healthy, errored, abnormal]);
+      expect(sessions.stabilityScore(), isNull);
 
-      session = Sessions([], [errored, abnormal, crashed]);
-      expect(session.stabilityScore(), isNull);
+      sessions = Sessions([], [errored, abnormal, crashed]);
+      expect(sessions.stabilityScore(), isNull);
     });
 
     test('stability score', () {
@@ -34,14 +34,18 @@ void main() {
       final abnormal = _givenSessionGroup(SessionStatus.abnormal, 1);
       final crashed = _givenSessionGroup(SessionStatus.crashed, 1);
 
-      var session = Sessions([], [healthy]);
-      expect(session.stabilityScore(), isNull);
+      final sessions = Sessions([], [healthy, errored, abnormal, crashed]);
+      expect(sessions.stabilityScore(), equals(75.0));
+    });
 
-      session = Sessions([], [healthy, errored]);
-      expect(session.stabilityScore(), isNull);
+    test('all zero equals 100', () {
+      final healthy = _givenSessionGroup(SessionStatus.healthy, 0);
+      final errored = _givenSessionGroup(SessionStatus.errored, 0);
+      final abnormal = _givenSessionGroup(SessionStatus.abnormal, 0);
+      final crashed = _givenSessionGroup(SessionStatus.crashed, 0);
 
-      session = Sessions([], [healthy, errored, abnormal, crashed]);
-      expect(session.stabilityScore(), equals(75.0));
+      final sessions = Sessions([], [healthy, errored, abnormal, crashed]);
+      expect(sessions.stabilityScore(), equals(100.0));
     });
 
   });
